DirectionalLight now depends on Transform.
I came to the conclusion that there's very little reason to keep a direction value on the directionalLight component. It's easier to instead use a Transform component and use its rotation to reprensent the direction. How much of a memory cost this will result in is something we'll find out.