Improved organization of Renderables.
All the different Renderable types now exist on the same level. Each renderable type is a different struct, removing the need for a type enumeration.

Only one skybox renderable can be active in a collection at a time.

This was mostly cleanup to make the interface more pleasant to interact with.