The same gizmo type can be used multiple times in a frame.
This does have a bug. If the order of calls for running multiple instances of a certain gizmo type changes, so will the order of the gizmo instances used at the call sights. I don't care, though, because I haven't found a reason to yet.

This also cleans up the gizmos a bit more. Calling them is a bit more friendly and the materials used by each gizmo type are now created once and shared between all instances.