Reintroduced the line mesh generator.
It was removed in 0167a3d because I was under the impression that it should stay in the project it was being used for. It turns out that it already has a use case in another project.

I don't necessarily want to share every little thing I make that is usable across every project, though. I might introduce another repository for Varkor extensions that will handle this in the future.

Even though it is not a component, I put the LineMesh in the comp directory because I am hoping to convert it into one.