Mesh::mIndexCount is now public
Being able to change how many indices of a mesh are drawn is important when dealing with custom geometry. I currently have a use case where I don't want to draw the entire element buffer, but I also don't want to waste the time of deleting it and creating a new allocation.