Ds::RbTree::Insert/Replace don't crash on failure.
It's the calling code's responsibility to decide whether to crash or not. I have a use case where I want to attempt an insertion and continue silently if it fails. I have another use case where I want a crash. That's why Insert and Replace now return a Result so the calling code can decide what to do.