Allow the use of the scope resolution operator when registering components.
It is not possible to use "::" in component names because ':' is used for the start and end of keys in Valkor and comonent names are used as keys. This poses a problem for registering components that are contained in namespaces, hence, I replace the scope resolution operator with '/' for component names.

There are some other possible solutions. One involves putting `using <namespace_name>` before registering the components. This means that the scope resolution operator doesn't need to be used for components in <namespace_name>. Though I have used this technique for some time, I don't like that because it ignores the exact reason why namespaces exist; avoiding name conflicts. I now have a situation where I am using a namespace other than `Comp::` and I find putting the namespace name in the component name to be important for clarity reasons.

The other possibility was to modify the Valkor tokenizer to accept keys containing ':'. This was undesirable because it would make it unclear when a key ends to the tokenizer. The lack of clarity would result in requiring a rule like "spaces after keys", in order to accept ':' in key identifiers. This would be a negative for Valkor because it would make what are currently simple rules, complicated.