Box Intersection and Unit Test Viewer
The concepts fueling the implementation of the BoxIntersection test come from Real-Time Collision Detection by Christer Ericson. I did not use the implementation given in the book. Though it's certainly faster, I did not understand all of the terms and opted for a implementation that I fully understand.

The unit test for box intersection does not cover all the possible cases. This should be addressed at some point. I have yet to see a situation where the box intersection results in an undersirable output.

The BoxIntersectionTest struct uses a member for the test name so the code documents which test is which when emplaced into the vector of BoxIntersectionTest instances. It will make easier to find the parameters used for a failing test when there happens to be one in the future. I plan on going forward with this structure for other unit test types.

The Unit Test Viewer was added to provide a means of displaying unit tests. Most unit tests don't need this, so it's something that unit tests can opt into using. The Viewer was added because it will be an extremely valueable resource for debugging all of the intersection functions and other physics related code that I need to write. Given the very visual nature of these operations, making a unit test viewer was a logical step.

The inclusion of the unit test viewer required different editor levels. The unit test viewer does not need the entire Varkor interface, so the introduction of a Simple editor level allows me to ignore the Varkor interface and make a custom small interface for interacting with the unit test viewer.