Sphere Triangle Intersection and Improved Unit Test viewer.
Sphere Triangle intersection boils down to finding the closest point to the center of the sphere that lies on the relevant triangle. If that point is within the sphere, there is an intersection.

The improvements to the unit test viewer just make it easier to extend it with different types of tests. There is no longer a need to add cases to switch statements for getting the number of tests under the test type, the specific name of a test, displaying the test, or displaying  the test's gizmos.