Rename mA, mB, and mC to mPoints within Math::Triangle
Well, firstly, this is my first commit in a long time, thus it's not super reliable. Secondly, this was an uncommitted change that's been sitting in my local copy of the repo for probably close to a year now. I don't think I had a reason other than it feels cleaner for it to be this way. That seems true becuase now it's nicer to refer to triangle points with a, b, and c instead of the previous member names.