Rename StandardPosition/StandardTranslation to NdcPosition.
It is well known what Ndc means and using that explains more clearly what the functions do. I decided to limit the use of NdcMousePosition to Input::NdcMousePosition, because it's the only function that relates specifically to the mouse. The other functions only take and recieve Vec2s, but those values don't necessarily need to be the mouse position.