Custom keybinds
I have a custom keyboard now and I don't find the keybinds I want to use to be sensible defaults. Since having the ability to change keybinds is a very necessary feature, not only for an engine, but for a game, it was a good time to implement custom keybinds. This should also serve as a nice reference for the future when I need to do this for games.

The line for preventing camera motion when the control modifier key is down was removed because the keybind modifier test prevents an action from being active if the control key is held down and the action does not expect the control key to be held down.