Member name's are now stored in a component.
Not every member needs a name. It is expected that a lot of members will be created programmatically and therefore never require a name. Not only that, but the Space should not be managing something like the name. It should only manage the members and their components, not any other data specific to the members.