Relationships are now stored in a component.
Not every member needs to keep track of relationships. It might often be the case that most members will not have any relationship. Putting this information on a component means that memory will not when there is no relationship information that needs to be tracked. This also improves the quality of save files because the relationship data is only included within the Components pair array when there is actually a relationship that needs to be tracked.

This is also an effort to decouple relationship data from the space. The purpose of the space is to track entities and their components. Having it also track relationship data muddies the responsibility of the Space. Unfortunately, relationship implementation is still closely tied to the Space, but decopling the implementation was beyond the scope of this work.