Improving world_Space test output
The member pool and the descriptor bin are no longer printed out during the space unit tests. There are two reasons for this.

One: Whenever a change was made to the implemenation of the space or the table, verifying the validity of the changes that occurred in the test output was incredibly taxing and didn't seem to be a productive use of time. It didn't seem productive because implementation details were being tested rather than focusing on testing expected functionality. The new output tries to focus on this functionality rather than how that functionality is achieved.

Two: The member pool and descriptor bin will be removed in a later commit due to a structural change in the implementation of space. By changing the test output in this commit, expect the output to remain the same despite the upcoming refactor.