Converted serialized spaces to pair arrays.
This change was only possible due to the fact that member names and member relationships were put into components in previous commits. This means members have no extra metadata. Because of that, we can make the serialized value of a space more pleasant to look at. We can use member ids as keys within a pair array and have the components pair array be the value of one of those pairs. This gets rid of the ugly value array of pair arrays that existed prior. Take a look at the world_Space unit test output for an example.