Bug fix for SparseSet::Request
Request originally had the id take ownership of the next available dense element, but it did not update that dense element to reference the requested id. It instead falsely pushed the id to the dense vector without considering that the dense usage might be less than the size of the dense vector.

Addressing this bug required removing the notion of invalid dense index references (the poorly named nInvalidSparseId) within the sparse vector. If we did not always track the map between sparse elements and dense elements, we would need to perform a linear search during requests. This search would be performed on the unused section of the dense vector in order to find the dense element that contains the requested id. We need to know where this dense element is because it needs to be swapped with the first element in the unused section. By doing so, the reference to the requested id will be in the used section and the reference to the still unused id will be in the unused section by the end of the request. By instead having every element of the sparse vector reference its associated dense element, finding the dense element that refers to a sparse element is just a matter of looking at the reference contained at the requested id within the sparse vector. This means request can remain a constant time operation while also maintaining that all ids in the used section of the dense vector are indeed in use and all ids in the unused section are indeed not in use.

It's also worth mentioning that removing the notion of invalid dense indices results in no loss of functionality. We can still determine that an id is used by checking if that id's dense element is less than the dense usage.