Tables use member ids to access components
This evolves the space into what one could consider a true ecs. The descriptor bin and member pool have been eliminated. This was the main intent behind the change. This means we are no longer perpetually growing the descriptor bin. This also means that GetComponent is now a constant time operation since we no longer need to search through a member's component descriptors to find the index of the desired component.

There are some other cool benefits. Tables no longer allow using the same member id twice and components are always serialized in the same order.

There are some cons too. Duplicating, deleting, and serializing a member are all slower because we need to search through all of the component tables to find the components owned by that member. These cons would need to be thoroughly tested to understand just much they cost. It will only become more of a problem as the number of components increases.