Space::DeleteMember bug fix.
If deleting a parent member caused the its relationship component to be moved before all of its children were deleted, some of the parent's children would not be deleted because the pointer to the parent's relationship component was invalidated. This invalidation occurred because it was moved to fill in a gap within the relationship component table that was created when a child's relationship component was deleted.

This also includes a test that forces this condition to occur.