Test for measuring performance of Space
The important test is AddComponentsDeleteMembers. This test exemplifies the problem solved by the previous refactor of Space.

'option' was changed to 'set' in ext/CMakeLists.txt because using option would not overwrite the previously cached values. Setting these options as such also seems valuable for testing the Varkor executable directly, but only a test will tell if that's truly the case.

The cmake code for creating the perf_space target is not generalized for other performance tests because there aren't any others. This should only be generalized when more are added.