Removed the usage of Ds::Vector from SparseSet
This wasn't entirely necessary, but using Ds::Vector didn't seem entirely appropriate for the allocations of the dense and sparse arrays. The full capacity of the Vectors were always used, therefore nullifying the use of the vectors' size. Additionally, two allocations were used when a single allocation could be used for everything needed by Ds::SparseSet.