Store tables in a Pool
This is certainly a form of preoptimization that isn't entirely grounded. Essentially, a map makes a separate allocation for every element added to the set, whereas a Pool makes one allocation to store everything in the set. By storing the Tables in a Pool, we get rid of these separate allocations for every set element. Pool also fits the exact use case of Space Tables since they are accessed using integer ids. Unfortunately, no noticeable performance benefits came from this.