btest.js integrated into build.js
This makes it possible to run all varkor utilities, projects, and tests using a single command. This removes a reasonable amount of friction experienced when running unit tests with the old btest.js script. Now tests take advantage of the fact that the build directory is saved and never has to be retyped on the command line. This is also true for the currently active test target.

It would have likely been possible to keep btest.js as its own top level script and somehow integrate some of the option saving functionality into it. The approach of merging the functionality into build.js was taken because it seemed easier.