Build script improvements
Yes/No toggling has been removed. This was annoying because up arrow enter would result in the behaviour switching back and forth.

The test action is no also saved to file so it doesn't need to be retyped every time. I have some fear that this behaviour won't play well with o(verwrite), but it's better to find out rather than speculate.