Static Quaternion Initializers
The Identity, AngleAxis, FromTo, and BasisVectors Quaternion member functions have been turned into static member functions that return a Quat. This is dangerous. Any code calling these functions on a Quat instance will not raise an error. The problem will show up as some sort of runtime bug.

```
Quat q;
// No error despite FromTo being a static member function
q.FromTo(from, to);
```

This change is still worthwhile since it improves the usability and expressability of the quaternion interface. It has resulted in the removal of multiple unnecessary lines of code.