Component Progression Bugfix
This should have been part of cbbbeda. Components are now part of a member's pair array rather than being stored in a "Components" pair array within the member pair array. I'm surprised this went undetected for a so long and I'm not quite sure why things worked without this change.