Gizmo handles update based on camera translation
Unfortunately, a bit of code is duplicated between the Translator and Scalor. This form of duplication has repeatedly occured between these constructs because of the number of similarities they share, but they are also different enough that generalization is difficult and/or uneccessary.

I made OrientHandlesTowardsCamera into a template function as a small test in an effort to reduce this duplictation. The implementation of the function and the extra includes within the header were cons that outweighed the minor amount of code duplication. Whether that's right or wrong, idk, but to question it further is a waste of time.