Initial Convex Hull
An initial incomplete implementation of quickhull. This implementation works for the point sets defined in the quickhull tests, but not for every point set. It does not use a dynamic epsilon, does not account for edge cases introduced by not merging faces, and does not use Newell planes for faces with more than 3 vertices (introduced through face merging).

The performance of the implementation can likely also be improved dramatically by addressing the multiple linear searches that are performed (searching for an already visited face during horizon finding and adding dead vertices by checking if the vertex is on the horizon or already in the set of dead vertices) and by using Pools instead of linked lists to store hull data.

The commit is not perfect, but a necessary historical timepoint.