Ds::HashMap
This is the real reason I made HashSet.

The unit test here isn't very imporant. I was just using it to make sure the Ds::HashMap calls worked as expected. For example, it was necessary to write Ds::Map::Remove, Find, and Contains separately because it wouldn't have otherwise been possible to use `const char*` to find `std::string`s in the HashMap.

I don't like the fact that ds/Hash.h includes the string header. It's not possible to get rid of this, so replacing it will involve finally replacing std::string. Who knows when and if that will happen.