Ds::HashSet::Grow fix
The original had major problem. It created a new HashSet and assigned that new HashSet to this. This ended up overwriting the value of mSize that was incremented beforehand. Putting the increment of mSize after the call to TryGrow() seemed like smell, hence fixing the Grow function became the action to take.