FaceConflictList HashMap
A long time ago, I thought it would be a good idea to store all of the face conflict lists in a HashMap. By doing so, finding the conflict list for a face becomes a constant time operation. Unfortunately, this makes the function slower in practice. I have not tested it rigorously enough to know where this slow down occurs, but my hunch is that a lack of cache coherency results in the slow down.

The real solution to the look-up problem would is probably to store the conflict lists as part of faces directly. Dirk Gregorious has explicitly mentioned this. I've been resistant to doing so because these conflict lists are only relevant to the QuickHull function and I don't want them to stick around after the fact. Maybe the proper solution is to have a struct for faces meant specifically for use in QuickHull that are then transfered to a different struct without the ConflictList member.

So much effort and no reward. But hey, now I finally have HashSet and HashMap.