QuickHull Test Vertex Sort
The previous changes to QuickHull resulted in the order of the vertices being different on each run of the QuickHull function. I'm not sure exactly why, but I'm strongly under the impression it has to do with using the addresses of Ds::List members as HashMap keys.

Ordering the vertices creates makes the output of QuickHull tests consistent, no matter the order they are stored in within the half edge structure.