QuickHull face merging improvements
I now ensure that only merges considered convex are perfomed. In order to determine whether a convex edge is a viable merge, I use the angle between the normals of the adjacent faces. I also now use an epsilon that is calculated from the span of the point set.

Other changes that were made address the crashes that arose due to the new merging rules. The quick hull code has also been cleaned up a bit (paragraphing and comments above unclear sections for better readability).