Math::Vector(VectorN-1, value) initialization
Having this constructor makes some code in a later commit much nicer.

I avoided putting constructors in my vector class for the longest time because I wanted to be able to directly initialize with an initialization list. This change unfortunately removes the ability to make constexpr vectors. This problem is the reason for a majority of changes.